# src/product/Gateway/Microsoft.Azure.EngagementFabric.RequestListener/StaticContent/Swagger/2018-10-01.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "SMS and Social Login",
    "version" : "1.0.0",
    "title" : "CEF API Prod (SMS, OTP AND SOCIAL LOGIN)",
    "contact" : {
      "email" : "haoliuh@microsoft.com"
    }
  },
  "host" : "cef.chinacloudapi.cn",
...
  "paths" : {
...
    "/services/sms/signatures/{signature}" : {
      "get" : {
        "tags" : [ "SMS" ],
        "summary" : "Read a SMS signature.",
        "description" : "Read a SMS signature.",
        "parameters" : [ {
          "name" : "api-version",
          "in" : "query",
          "description" : "Version of API call.",
          "required" : true,
          "type" : "string",
          "default" : "2018-10-01"
        }, {
          "name" : "signature",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "default" : "Disney China"
        }, {
          "name" : "Account",
          "in" : "header",
          "description" : "Tenant account in CEF.",
          "required" : true,
          "type" : "string",
          "default" : "MockAccount"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "SASToken for authorization",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Return the signature",
            "schema" : {
              "$ref" : "#/definitions/smsSignature"
            }
          },
          "400" : {
            "description" : "Invalid request body. The registration could not be created because the request was malformed."
          },
          "401" : {
            "description" : "Authorization failure. The access key was incorrect."
          }
        }
      },
      "delete" : {
        "tags" : [ "SMS" ],
        "summary" : "Delete a sms signature.",
        "description" : "Delete a sms signature given a signature name.",
        "parameters" : [ {
          "name" : "api-version",
          "in" : "query",
          "description" : "Version of API call.",
          "required" : true,
          "type" : "string",
          "default" : "2018-10-01"
        }, {
          "name" : "signature",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "default" : "Disney China"
        }, {
          "name" : "Account",
          "in" : "header",
          "description" : "Tenant account in CEF.",
          "required" : true,
          "type" : "string",
          "default" : "MockAccount"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "SASToken for authorization",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Deleted"
          },
          "400" : {
            "description" : "Invalid request body. The registration could not be created because the request was malformed."
          },
          "401" : {
            "description" : "Authorization failure. The access key was incorrect."
          }
        }
      }
    },
...
    "/services/otp/start" : {
      "post" : {
        "tags" : [ "SMS (OTP)" ],
        "summary" : "Send verfification to end users.",
        "description" : "Send verfification to end users.",
        "parameters" : [ {
          "name" : "api-version",
          "in" : "query",
          "description" : "Version of API call.",
          "required" : true,
          "type" : "string",
          "default" : "2018-10-01"
        }, {
          "name" : "Account",
          "in" : "header",
          "description" : "Tenant account in CEF.",
          "required" : true,
          "type" : "string",
          "default" : "MockAccount"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "SASToken for authorization",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "otpstartrequest",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/otpstartrequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "operation is done successfully.",
            "schema" : {
              "type" : "object",
              "properties" : {
                "expireTime" : {
                  "type" : "integer",
                  "example" : 300
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request body. The registration could not be created because the request was malformed."
          },
          "401" : {
            "description" : "Authorization failure. The access key was incorrect."
          }
        }
      }
    },
    "/services/otp/check" : {
      "post" : {
        "tags" : [ "SMS (OTP)" ],
        "summary" : "Verify code.",
        "description" : "Verify code",
        "parameters" : [ {
          "name" : "api-version",
          "in" : "query",
          "description" : "Version of API call.",
          "required" : true,
          "type" : "string",
          "default" : "2018-10-01"
        }, {
          "name" : "Account",
          "in" : "header",
          "description" : "Tenant account in CEF.",
          "required" : true,
          "type" : "string",
          "default" : "MockAccount"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "SASToken for authorization",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "otpcheckrequest",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/otpcheckrequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "operation is done sucessfully.",
            "schema" : {
              "type" : "object",
              "properties" : {
                "state" : {
                  "type" : "string",
                  "example" : "SUCCESS"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request body. The registration could not be created because the request was malformed."
          },
          "401" : {
            "description" : "Authorization failure. The access key was incorrect."
          }
        }
      }
    }
  },
  "definitions" : {
...
    "smsSignature" : {
      "type" : "object",
      "properties" : {
        "signature" : {
          "type" : "string",
          "example" : "Disney(China)"
        },
        "state" : {
          "type" : "string",
          "example" : "Pending"
        },
        "stateMessage" : {
          "type" : "string",
          "example" : "",
          "description" : "The reason why the approvement is failed. Only applicable when state is failed. Input by CEF operation."
        }
      },
      "example" : {
        "signature" : "Disney(China)",
        "stateMessage" : "",
        "state" : "Pending"
      }
    },
...
    "otpstartrequest" : {
      "type" : "object",
      "required" : [ "phoneNumber", "templateName" ],
      "properties" : {
        "phoneNumber" : {
          "type" : "string",
          "example" : "1892291902"
        },
        "templateName" : {
          "type" : "string",
          "example" : "OTPTemplate",
          "description" : "The template must contain CEF-reserved param \"$otpcode\". e.g. \"【xxx】你的验证码是\" + $(otpcode) + \", 请在五分钟内输入。\""
        },
        "expireTime" : {
          "type" : "integer",
          "example" : 300,
          "description" : "Seconds to expire. If users make another request within the xx minutes, they will receive the same token. If not specified the default value is 300."
        },
        "codeLength" : {
          "type" : "integer",
          "example" : 6,
          "description" : "The length of the code min 4~ max 8. the default is 6."
        },
        "channel" : {
          "type" : "string",
          "example" : "sms",
          "description" : "The channel (\"sms\", \"call\", \"email\") to send the code. The default is sms."
        }
      }
    },
...
    "otpcheckrequest" : {
      "type" : "object",
      "properties" : {
        "phoneNumber" : {
          "type" : "string",
          "example" : "1892291902"
        },
        "channel" : {
          "type" : "string",
          "example" : "sms"
        },
        "code" : {
          "type" : "string",
          "example" : "310928"
        }
      }
    },
...
  },
...
}